AG-9002 Move applyDeprecations() to rangeArea.ts
