+++ packages/ag-charts-enterprise/src/series/range-area/rangeArea.ts
@@ -8,7 +8,7 @@ import {
     type AgSeriesMarkerStyle,
     _ModuleSupport,
 } from 'ag-charts-community';
-import type { DeepRequired, Point, RequireOptional } from 'ag-charts-core';
+import type { AreExact, DeepRequired, Point, RequireOptional } from 'ag-charts-core';
 
 import { type RangeAreaSeriesDataAggregationFilter, aggregateRangeAreaData } from './rangeAreaAggregation';
 import { calculateIntersectionSegments, findRangeAreaIntersections } from './rangeAreaIntersection';
@@ -501,8 +501,124 @@ export class RangeAreaSeries extends _ModuleSupport.CartesianSeries<
         return low.marker.isDirty() || high.marker.isDirty();
     }
 
+    public applyDeprecations() {
+        // We could, in theory, use the graph config to apply the deprecations.
+        //
+        // Example:
+        //    { item: { low: { strokeWidth: { $path: [ '../../strokeWidth', 1 ] }
+        //
+        // But we can only use $path and $isUserOption on leaf-nodes. So in practice, things get overly complex when
+        // trying to configure the defaults for `item.[low|high].marker.fill`, because the defaults also depend on
+        // `fill.type`.
+        //
+        // So simplify this by implementing the deprecated options here.
+        //
+        const {
+            lineDash: deprecatedLineDash,
+            lineDashOffset: deprecatedLineDashOffset,
+            stroke: deprecatedStroke,
+            strokeOpacity: deprecatedStrokeOpacity,
+            strokeWidth: deprecatedStrokeWidth,
+            item: { low, high },
+        } = this.properties;
+
+        if (deprecatedLineDash != null) {
+            low.lineDash = deprecatedLineDash;
+            high.lineDash = deprecatedLineDash;
+        }
+        if (deprecatedLineDashOffset != null) {
+            low.lineDashOffset = deprecatedLineDashOffset;
+            high.lineDashOffset = deprecatedLineDashOffset;
+        }
+        if (deprecatedStroke != null) {
+            low.stroke = deprecatedStroke;
+            high.stroke = deprecatedStroke;
+        }
+        if (deprecatedStrokeOpacity != null) {
+            low.strokeOpacity = deprecatedStrokeOpacity;
+            high.strokeOpacity = deprecatedStrokeOpacity;
+        }
+        if (deprecatedStrokeWidth != null) {
+            low.strokeWidth = deprecatedStrokeWidth;
+            high.strokeWidth = deprecatedStrokeWidth;
+        }
+
+        const deprecatedMarker = this.properties.marker == null ? undefined : { ...this.properties.marker };
+        if (deprecatedMarker != null) {
+            deprecatedMarker.enabled ??= true; // auto-enable
+            const {
+                enabled: deprecatedMarkerEnabled,
+                stroke: deprecatedMarkerStroke,
+                itemStyler: deprecatedItemStyler,
+                size: deprecatedSize,
+                shape: deprecatedShape,
+                fill: deprecatedFill,
+                fillOpacity: deprecatedFillOpacity,
+                strokeOpacity: deprecatedStrokeOpacity,
+                strokeWidth: deprecatedStrokeWidth,
+                lineDash: deprecatedLineDash,
+                lineDashOffset: deprecatedLineDashOffset,
+                ...rest
+            } = deprecatedMarker;
+
+            // check that all properties from deprecatedMarker have been exhausted:
+            true satisfies AreExact<typeof rest, {}>;
+
+            if (deprecatedMarkerEnabled != null) {
+                low.marker.enabled = deprecatedMarkerEnabled;
+                high.marker.enabled = deprecatedMarkerEnabled;
+            }
+            if (deprecatedMarkerStroke != null) {
+                low.marker.stroke = deprecatedMarkerStroke;
+                high.marker.stroke = deprecatedMarkerStroke;
+            }
+            if (deprecatedItemStyler != null) {
+                low.marker.itemStyler = deprecatedItemStyler;
+                high.marker.itemStyler = deprecatedItemStyler;
+            }
+            if (deprecatedSize != null) {
+                low.marker.size = deprecatedSize;
+                high.marker.size = deprecatedSize;
+            }
+            if (deprecatedShape != null) {
+                low.marker.shape = deprecatedShape;
+                high.marker.shape = deprecatedShape;
+            }
+            if (deprecatedFill != null) {
+                if (typeof deprecatedFill === 'string') {
+                    low.marker.fill = deprecatedFill;
+                    high.marker.fill = deprecatedFill;
+                } else {
+                    // TODO: gradient/pattern/image
+                }
+                low.marker.fill = deprecatedFill;
+                high.marker.fill = deprecatedFill;
+            }
+            if (deprecatedFillOpacity != null) {
+                low.marker.fillOpacity = deprecatedFillOpacity;
+                high.marker.fillOpacity = deprecatedFillOpacity;
+            }
+            if (deprecatedStrokeOpacity != null) {
+                low.marker.strokeOpacity = deprecatedStrokeOpacity;
+                high.marker.strokeOpacity = deprecatedStrokeOpacity;
+            }
+            if (deprecatedStrokeWidth != null) {
+                low.marker.strokeWidth = deprecatedStrokeWidth;
+                high.marker.strokeWidth = deprecatedStrokeWidth;
+            }
+            if (deprecatedLineDash != null) {
+                low.marker.lineDash = deprecatedLineDash;
+                high.marker.lineDash = deprecatedLineDash;
+            }
+            if (deprecatedLineDashOffset != null) {
+                low.marker.lineDashOffset = deprecatedLineDashOffset;
+                high.marker.lineDashOffset = deprecatedLineDashOffset;
+            }
+        }
+    }
+
     override update(a: Parameters<BaseUpdate>[0]): ReturnType<BaseUpdate> {
-        this.properties.applyDeprecations();
+        this.applyDeprecations();
         return super.update(a);
     }
